requirements.txt
# SAFE SCORING - Dependencies

# Core
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0

# Web Interface
flask>=3.0.0

# Scraping (optional)
playwright>=1.40.0

# Utils
python-dotenv>=1.0.0

# Proprietary Data Collection (creates competitive moat)
schedule>=1.2.0          # Job scheduling for automated collection
tweepy>=4.14.0           # Twitter API (optional - for incident scraping)
aiohttp>=3.9.0           # Async HTTP for faster scraping
